# backend/openapi/dist.yaml
openapi: 3.0.3
info:
  title: Modular JWT Authenticated API
  version: 1.0.0
  description: An example API using JWT imported from a separate file.
servers:
  - url: 'http://localhost:8080/api/v1'
security:
  - BearerAuth: []
paths:
...
  /events:
    post:
      summary: Create a new event
      operationId: createEvent
      tags:
        - Event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_CreateEvent'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events/list:
    get:
      summary: List all events for the current user
      operationId: listEvents
      tags:
        - Event
      parameters:
        - name: type
          in: query
          description: Filter by event type
          schema:
            type: string
            enum:
              - task
              - event
              - note
        - name: status
          in: query
          description: Filter by status (for tasks)
          schema:
            type: string
            enum:
              - pending
              - completed
              - cancelled
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Res_DetailEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
  '/events/{id}':
    get:
      summary: Get a specific event by ID
      operationId: getEventById
      tags:
        - Event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/events/{id}/update':
    patch:
      summary: Update an event
      operationId: updateEvent
      tags:
        - Event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_UpdateEvent'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/events/{id}/delete':
    delete:
      summary: Delete an event
      operationId: deleteEvent
      tags:
        - Event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Event deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /events/range:
    get:
      summary: List events within a date/time range
      operationId: listEventsByRange
      tags:
        - Event
      parameters:
        - name: start_time
          in: query
          required: true
          description: Start date/time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          description: End date/time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: type
          in: query
          description: Filter by event type
          schema:
            type: string
            enum:
              - task
              - event
              - note
      responses:
        '200':
          description: List of events in range
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Res_DetailEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: Bearer <token>.
        The token can be obtained from the `/auth/login` endpoint.
    EditToken:
      type: apiKey
      in: header
      name: X-Edit-Token
      description: Public edit token for note collaboration
  schemas:
...
    Req_CreateEvent:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        type:
          type: string
          enum:
            - task
            - event
            - note
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        due_date:
          type: string
          format: date
        status:
          type: string
          enum:
            - pending
            - completed
            - cancelled
        priority:
          type: string
          enum:
            - low
            - normal
            - high
        category_id:
          type: integer
        is_all_day:
          type: boolean
          default: false
        google_event_id:
          type: string
          nullable: true
        source:
          type: string
          default: local
      required:
        - title
        - type
        - start_time
    Req_UpdateEvent:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        due_date:
          type: string
          format: date
        status:
          type: string
          enum:
            - pending
            - completed
            - cancelled
        priority:
          type: string
          enum:
            - low
            - normal
            - high
        category_id:
          type: integer
        is_all_day:
          type: boolean
...
    Res_DetailEvent:
      type: object
      properties:
        id:
          type: string
          description: Unique event identifier
        user_id:
          type: integer
          description: User who created the event
        title:
          type: string
          description: Event title
        description:
          type: string
          description: Event description
        tags:
          type: array
          items:
            type: string
          description: Event tags
        type:
          type: string
          enum:
            - task
            - event
            - note
          description: Event type
        start_time:
          type: string
          format: date-time
          description: Event start time
        end_time:
          type: string
          format: date-time
          description: Event end time (optional)
        due_date:
          type: string
          format: date
          description: Due date (mainly for tasks)
        status:
          type: string
          enum:
            - pending
            - completed
            - cancelled
          description: Status (only for tasks)
        priority:
          type: string
          enum:
            - low
            - normal
            - high
          description: Priority (only for tasks)
        category_id:
          type: integer
          description: Category ID
        is_all_day:
          type: boolean
          default: false
          description: Whether event is all-day
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        google_event_id:
          type: string
          nullable: true
          description: Google Event ID if synced
        source:
          type: string
          description: Event source (local or google)
          default: local
      required:
        - id
        - user_id
        - title
        - type
        - start_time
...
  responses:
    BadRequest:
      description: Bad Request. The request is invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request.
    Unauthorized:
      description: 'Unauthorized. The provided JWT token is missing, invalid, or expired.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication token is required.
    NotFound:
      description: Not Found. The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found.
...